fix a permissions issue on the server
basically everything about becoming an unprivileged user is pain.
just fix the permissions after doing what you need to do
as root. so this change basically pulls in the git repo
installs the virtualenv (as root) and then changes
the owner/group on all of that to be the bot.
SEEMS LIKE A LOT OF WORK ANSIBLE